requirements.txt
pandas>=2.0.0
requests>=2.31.0
python-dateutil>=2.8.2
icalendar>=5.0.7
schedule>=1.2.1
openpyxl>=3.1.2